Build method for handling cookie creation
